# engine: skip WAL for temp rocks store