The "temp" engine is used by some SQL processors and IMPORT to buffer
data that is too large to fit in memory. It is deleted and recreated on
startup, so writing to the WAL is pointless overhead.

Importing the TPC-H lineitem table at scale-factor 10 (~7.2GB in CSV) on
an otherwise unused, uncontended 4 node cluster, this cut the averge
completion time from 31m15s to 27m52s. Gains are likely more pronounced
in the presense of more IPOS or disk bandwidth contention.

RocksDB is still likely overkill for our usage here and additional gains
seem likely by replacing it entirely, but this seems like a quick win
for now.

Release note: None