@jordanlewis 2018-12-14
Benchmark results for the processors that use disk? There are a few that specifically measure disk spilled usage that would be good to see.