@dt 2018-12-14
I mostly just wanted to get this looked at by those with more rocks expertise, in case I was doing something stupid or missing something, before I put too much time into benchmarking something with an obvious mistake. 

I don't see any difference in the `hashjoiner/spill=true` benchmark -- I'm guessing it is doing enough other work that it doesn't bottleneck on the extra disk IO? I tried it on a slower-disk cloud worker as well as my faster-disk MBP with different benchtimes and didn't see much of interest on either.

For the IMPORT measurements, I did several runs of TPC-H 10 lineitem and then just pulled all the times from `show jobs` on each cluster.  I was using two side-by-side clusters, rather than the same cluster and swapping binaries back and forth, so possible I could have had noisy neighbors or something. Here's the timings from the jobs table:

master | no_wal
-- | --
0:34:41 | 0:33:11
0:29:17 | 0:32:06
0:31:22 | 0:30:13
0:30:50 | 0:30:23
0:34:31 | 0:29:27
0:32:13 | 0:27:57
0:32:09 | 0:27:09
0:28:22 | 0:28:12
0:29:00 | 0:26:38
Average  |  
0:31:23 | 0:29:28