@a-robinson 2018-12-17
To back up @petermattis, I did some tests on this back when I was working on node liveness stalls, and I also wasn't able to improve performance at all by disabling the WAL.