@petermattis 2018-12-17
I've been futzing with the pebble version some more and have newer/better numbers:

```
~ benchstat out.rocksdb out.pebble
name                                    old time/op  new time/op  delta
RocksDBMapWrite/InputSize4096-8         5.97ms ±13%  1.34ms ± 5%  -77.57%  (p=0.000 n=20+20)
RocksDBMapWrite/InputSize16384-8        15.9ms ± 4%   5.5ms ± 1%  -65.38%  (p=0.000 n=19+18)
RocksDBMapWrite/InputSize65536-8        57.8ms ± 1%  24.1ms ± 1%  -58.32%  (p=0.000 n=19+18)
RocksDBMapWrite/InputSize262144-8        257ms ± 2%   108ms ± 2%  -58.06%  (p=0.000 n=17+20)
RocksDBMapWrite/InputSize1048576-8       1.60s ± 4%   0.59s ± 4%  -63.23%  (p=0.000 n=20+19)
RocksDBMapIteration/InputSize4096-8     1.24ms ± 1%  0.26ms ± 2%  -79.42%  (p=0.000 n=19+20)
RocksDBMapIteration/InputSize16384-8    6.27ms ± 1%  1.13ms ± 1%  -81.96%  (p=0.000 n=17+19)
RocksDBMapIteration/InputSize65536-8    26.9ms ± 1%   4.7ms ± 3%  -82.62%  (p=0.000 n=19+18)
RocksDBMapIteration/InputSize262144-8    135ms ± 1%    40ms ± 2%  -70.71%  (p=0.000 n=17+19)
RocksDBMapIteration/InputSize1048576-8   584ms ± 1%   203ms ± 0%  -65.26%  (p=0.000 n=18+18)
```

One of the improvements here was to disable the WAL in pebble. I had originally tried that and didn't see it make a difference. I'm pretty sure that was because disabling the WAL is a small fixed benefit and it gets swamped by other work involved. It isn't until that other work is optimized that disabling the WAL is worthwhile.

My takeaway from this experimentation is that we can do disk-sorting ~3x faster than our current implementation. We should be looking for those kinds of gains.